Lenguajes de Programación first exam
These two output the exact same stuff to the terminal, and do the math similarly. This is supposed to illustrate how all programming languages are replaceable with each other. 

examen.c can be compiled with any GNU-compliant compiler for the C programming language.
examen.js can theoretically be executed by any JavaScript interpreter, but it was written with Rhino in mind, to output to the terminal.